+++ FRAuth/FRAuth/DeviceBinding/DeviceBindingAuthenticators.swift
@@ -30,6 +30,13 @@ public protocol DeviceAuthenticator {
     /// - Returns: compact serialized jws
     func sign(keyPair: KeyPair, kid: String, userId: String, challenge: String, expiration: Date) throws -> String
     
+    /// Sign the challenge sent from the server and generate signed JWT
+    /// - Parameter userKey: user Information
+    /// - Parameter challenge: challenge received from server
+    /// - Parameter expiration: experation Date of jws
+    /// - Returns: compact serialized jws
+    func sign(userKey: UserKey, challenge: String, expiration: Date) throws -> String
+    
     /// Check if authentication is supported
     func isSupported() -> Bool
     
@@ -49,15 +56,14 @@ extension DeviceAuthenticator {
     /// - Parameter expiration: experation Date of jws
     /// - Returns: compact serialized jws
     func sign(keyPair: KeyPair, kid: String, userId: String, challenge: String, expiration: Date) throws -> String {
-        let jwk = try ECPublicKey(publicKey: keyPair.publicKey, additionalParameters: ["use": "sig", "alg": "ES256"])
-        let jwkWithKeyId = try jwk.withThumbprintAsKeyId()
+        let jwk = try ECPublicKey(publicKey: keyPair.publicKey, additionalParameters: [JWKParameter.keyUse.rawValue: "sig", JWKParameter.algorithm.rawValue: "ES256", JWKParameter.keyIdentifier.rawValue: kid])
         let algorithm = SignatureAlgorithm.ES256
         
         //create header
         var header = JWSHeader(algorithm: algorithm)
         header.kid = kid
         header.typ = "JWS"
-        header.jwkTyped = jwkWithKeyId
+        header.jwkTyped = jwk
         
         //create payload
         let params: [String: Any] = ["sub": userId, "challenge": challenge, "exp": (Int(expiration.timeIntervalSince1970))]
@@ -75,11 +81,44 @@ extension DeviceAuthenticator {
         return jws.compactSerializedString
     }
     
+    
+    // Default implemention
+    /// Sign the challenge sent from the server and generate signed JWT
+    /// - Parameter userKey: user Information
+    /// - Parameter challenge: challenge received from server
+    /// - Parameter expiration: experation Date of jws
+    /// - Returns: compact serialized jws
+    func sign(userKey: UserKey, challenge: String, expiration: Date) throws -> String {
+        guard let keyStoreKey = KeyAware.getSecureKey(keyAlias: userKey.keyAlias) else {
+            throw DeviceBindingStatus.unsupported(errorMessage: "Cannot read the private key")
+        }
+        let algorithm = SignatureAlgorithm.ES256
+        
+        //create header
+        var header = JWSHeader(algorithm: algorithm)
+        header.kid = userKey.kid
+        header.typ = "JWS"
+        
+        //create payload
+        let params: [String: Any] = ["sub": userKey.userId, "challenge": challenge, "exp": (Int(expiration.timeIntervalSince1970))]
+        let message = try JSONSerialization.data(withJSONObject: params, options: [])
+        let payload = Payload(message)
+        
+        //create signer
+        guard let signer = Signer(signingAlgorithm: algorithm, key: keyStoreKey) else {
+            throw DeviceBindingStatus.unsupported(errorMessage: "Cannot create a signer for jws")
+        }
+        
+        //create jws
+        let jws = try JWS(header: header, payload: payload, signer: signer)
+        
+        return jws.compactSerializedString
+    }
 }
 
 
 /// DeviceAuthenticator adoption for biometric only authentication
-internal struct BiometricOnly: DeviceAuthenticator {
+internal class BiometricOnly: DeviceAuthenticator {
     /// prompt description for authentication promp if applicable
     var promptDescription: String
     /// local authentication policy for authentication
@@ -133,7 +172,7 @@ internal struct BiometricOnly: DeviceAuthenticator {
 
 
 /// DeviceAuthenticator adoption for biometric and Device Credential authentication
-internal struct BiometricAndDeviceCredential: DeviceAuthenticator {
+internal class BiometricAndDeviceCredential: DeviceAuthenticator {
     /// prompt description for authentication promp if applicable
     var promptDescription: String
     /// local authentication policy for authentication
@@ -186,7 +225,7 @@ internal struct BiometricAndDeviceCredential: DeviceAuthenticator {
 }
 
 
-internal struct None: DeviceAuthenticator {
+internal class None: DeviceAuthenticator {
     
     /// keyAware for key pair generation
     var keyAware: KeyAware