+++ FRAuth/FRAuth/DeviceBinding/DeviceBindingStatus.swift
@@ -18,6 +18,7 @@ public enum DeviceBindingStatus: LocalizedError {
     case timeout
     case abort
     case unsupported(errorMessage: String?)
+    case unRegister
 }
 
 
@@ -33,6 +34,8 @@ public extension DeviceBindingStatus {
             return "Abort"
         case .unsupported:
             return "Unsupported"
+        case .unRegister:
+            return "Unsupported"
         }
     }
     
@@ -46,6 +49,8 @@ public extension DeviceBindingStatus {
             return "User Terminates the Authentication"
         case .unsupported(let errorMessage):
             return errorMessage ?? "Device not supported. Please verify the biometric or Pin settings"
+        case .unRegister:
+            return "PublicKey or PrivateKey Not found in Device"
         }
     }
 }
@@ -60,3 +65,14 @@ public enum DeviceBindingResult {
 
 /// Callback definition for completion of Device Binding
 public typealias DeviceBindingResultCallback = (_ result: DeviceBindingResult) -> Void
+
+
+/// Result enum for Device Signing
+public enum DeviceSigningResult {
+    case success
+    case failure(DeviceBindingStatus)
+}
+
+
+/// Callback definition for completion of Device Signing
+public typealias DeviceSigningResultCallback = (_ result: DeviceSigningResult) -> Void