+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 #### Added
 - Device Binding callback implementation
+- Device Signing Verifier callback implementation
 ## [3.4.0]
 #### Added
 - Dynamic SDK Configuration [SDKS-1760]